time works

+++ commands/help.go
@@ -5,7 +5,7 @@ import (
 	"github.com/eren5960/essentialsgo/global"
 	"math"
 )
-
+// Eren5960 <ahmederen123@gmail.com>
 type Help struct {
 	Page int `optional:""`
 }
@@ -25,6 +25,6 @@ func (t Help) Run(_ cmd.Source, output *cmd.Output) {
 	}
 
 	if output.MessageCount() == 1 {
-		output.Errorf("There are only have %d pages.", math.Round(float64(len(global.Commands)) / 5))
+		output.Errorf("There are only have %d pages.", int(math.Round(float64(len(global.Commands)) / 5)))
 	}
 }
